Extract parsing and handling logic to own classes
